Added license for Multiply() and Divide()
Thank you OpenZepplin for providing such a great library and helping to push forward the security of smart contracts in general. We are including your license here because we are combining your work with DappHub's for this library,  the apache license is a little more restrictive so we included that at the top and are posting yours here.